refactor 'ctd' and 'section' code (large changes)
This work was precipitated by a need to handle salinity error codes in
WOCE data.  It ended up involving many changes, to ctd code, e.g.
read.ctd(), as.ctd(), and summary.ctd(), as well as data(ctd), and to
section code (again, reading, creating, and data).  There is a
realistic chnce that this will cause problems, but I re-built
data(ctd), data(section), and data(a03), and all seemed OK, so this is
acceptable for a git push.