For short time-series plots, use hourly sub-ticks
Three-days plots were just showing one tick per day, which was not
ideal because users often want to see hours in plots of such short span.
Some adjustment may be needed in the categories, e.g. the code now
switches to 2-hour sub-ticks for a certain length.  I may use skynet as
a test case on this, because it always seems better to base methods on
actual data, which the eye tends to examine more critically than is the
case for made-up test cases.