for adp enu calc, try .C() instead of .Call()
It's still slower than the lapply method.  I really would like to
understand why, so I'll play with this some more.  Note that the .C()
interface is easier to code than the newer .Call() interface, and
there's no need to do R within the C, so that's the best.  Maybe I
should drop this into the ADV code, and see how it fares -- after all,
it's the time-consuming part.

> options(oce.flag1=0);source('~/src/R-kelley/oce/R/adp.R');system.time(d <- xyz.to.enu.adp(m05.adp.xyz, declination=(-18.099)))
oce.flag1 == 0 so using lapply and internal matrix operations
   user  system elapsed
  1.427   0.227   1.651

> options(oce.flag1=1);source('~/src/R-kelley/oce/R/adp.R');system.time(d <- xyz.to.enu.adp(m05.adp.xyz, declination=(-18.099)))
oce.flag1 == 1 so using .Call
   user  system elapsed
  5.599   2.708   8.222

> options(oce.flag1=2);source('~/src/R-kelley/oce/R/adp.R');system.time(d <- xyz.to.enu.adp(m05.adp.xyz, declination=(-18.099)))
oce.flag1 == 2 so using .C
   user  system elapsed
  5.732   4.252   9.918