working issue 116 (NOT DONE YET)
I dropped the Fortran() and .Call() methods, because they are no faster
than the .C() method.  I am seeing speed improvements on my laptop, but
not on my desktop, and I think the latter may have a general problem.
Note that the adp code still has remnants of the R method, and that
means doing the sin() and cos() work twice, even for the .C() method.
The next step will be to remove this duplication and clean the adp code.
Then the adv code will be updated, clearing these issues while they are
still on my mind.  I'm checking in partly to get the code to the desktop
machine, so I can try to figure out why the speeds are so odd there ...
if something is making my desktop 10 times slower than my laptop, I
would like to figure that out!