plot.adp(): return a list in some cases
For which="uv+ellipse", return information that can be used to construct
an ellipse separately.  For which="uv+ellipse+arrow", also return
mean-flow information.

The following sample code uses this sort of info (with the list set
manually for illustration); something like this  will be added to
man/plot.adp.Rd in due course.

> a <- list(ellipseMinor=1, ellipseMajor=2, ellipseAngle=25, meanU=1, meanV=2)
> y <- a$ellipseMajor * cos(t) * sin(-a$ellipseAngle) + a$ellipseMinor * sin(t) * cos(-a$ellipseAngle)
> x <- a$ellipseMajor * cos(t) * cos(-a$ellipseAngle) + a$ellipseMinor * sin(t) * sin(-a$ellipseAngle)
> plot(x, y, type='l', asp=1)
> arrows(0, 0, a$meanU, a$meanV)