prepare to decimate echosounder data
NOTE: I just discovered that decimate() does not work on ANY data.  I
had an early return of the original object.  But there is a lot of code
there, which tells me that I got partway to making this work and became
distracted and put in a workaround so R could check the package.  Bug
insertion by bug detection -- a nice touch for the lazy coder :-)