decimate(echosounder): start coding
Discusion.

In the example below, the image looks pretty good for practical
purposes, and the plot time scales linearly with total image size, so
it's much faster.

At the moment, I do a runmed before the bin avg.  If this proves useful,
I'll document it, but for now this is hidden.

Bugs.

1. the subdivisions must span exactly the original space
2. this is not documented.

Timing tests.

    d <- read.echosounder('~/Dropbox/clarkdanshared/echosounder/20080701_164323.dt4',despike=TRUE)

    system.time(plot(d))

       user  system elapsed
      2.979   0.544   6.439

    dd<-decimate(d, by=c(10,2))

    system.time(plot(dd))

    There were 50 or more warnings (use warnings() to see the first 50)
       user  system elapsed
      0.134   0.037   0.370