Improve plot.section() and read.ctd().
1. plot.section() now has a variety which="data", to indicate the
sampling locations.  This name is not very good, so the argument will
likely change ... and it's not really documented.

2. read.ctd() now handles better the case in which depth is found in a
seabird data file, but not pressure.  It does this by calculating the
whole-column median salinity and temperature (in hopes that the median
throws away crazy values like negative salinities) and then it computes
sigma-theta from that, and then employs a hydrostatic calculation to
compute a scale factor between depth and pressure.  Perhaps it should
also ensure that depth has the correct sign...