(re)fix issue198; start scale/resolution reporting
This commit is mainly about issue198.  The reporting part will involve
storing in the metadata the items

    velocityScale
    velocityResolution

etc.  I intend to do that for all quantities, and for all objects.  For
a start (since this is of present interest) it will be just velocity,
and just ADV objects.  The next step (just since it's easier to work
with code that is becoming familiar from other work) will be
temperature, etc. for ADV objects.  Then I'll do the same things for ADP
objects, where the resolution and scale are of more concern, since
storing a velocity in the 1-byte space used for ADP is fraught with
problems.