teos (broken)
I'm instrumenting a lot of code, to see what's wrong.  The test code
below exercises this.  I am starting to feel that I should just add a
teos argument to swRho(), and propagate results upward.  But actually
... maybe I should have a "eos" argument, with value 0 meaning UNESCO,
value 1 meaning TEOS-10, etc.  This could go in all the low-level
functions in sw.R and sw.c, and that way, once the tendrils that descend
from the highest-level functions connect with those that ascend from the
lower level ones, we'll be home free, with less case-by-case
conditionals at the high level.  Another advantage of this, beyond code
simplicity, is that it will permit users make their own EOS decisions,
at levels from top to bottom.  It will involve a fair bit of messing
about, because the changes will be broader than those in the present
scheme, but I don't mind a few days of code confusion (on this branch)
if the result is a code base that is robust, flexible, and maintainable.