teos(): speed up function calls (hugely)
Before this switch, the teos style for plotTS() took about 3 seconds.
The problem was that dlopen() was being called for single evaulation of
a TEOS function.  Drawing a TS diagram takes of order 1e4 function
calls.  Since the TEOS library is about 1e6 byte in size, that means
moving 1e10 bytes across the bus (or across a within-chip bus, depending
on the CPU cache size).  This means a transfer rate of roughly 3e9 bytes
per second.  I don't know the allowed speeds on my machines, but this
seems like it might be close to the speed of the devide.