echosounder work (incomplete)
Now decoding quite a lot more of the header information, with the
eventual goal of converting amplitude from counts to dB.  This involves
a lot of cross checking in the docs and with the strings in the
biosonics 01-Fish.dt4 sample file.  My hope, as I work through the
manual, is to gather better understanding of how to decode the
profiles, which I think is wrong on run-length decoding and also
(perhaps) in something more major, given the mismatch to graphs
produced with biosonics software on some of the three sample files.  (I
still am perplexed by how the depth-time patterns can look so alike in
one sample file and unalike in another.)