plot.ctd() now stops if 'xlim' or 'ylim' given, and explains why
I noticed that plot.lm() had the same sort of error as plot.ctd() was
giving, so I suppose it was OK, but I'd like oce to give better hints,
so I'll probably add code like this to e.g. plot.tdr() and other
things.  Just did ctd first because that's the filename that my fingers
are most used to typing.