binApply2D() speedup; make args similar to binMean2D()
For my test data (secchi) the speedup is a factor of about 100,
corresponding to the width of the matrix.  I think the issue is in
filling up the matrix, which is now a row at a time.  If this is still
slow, there are other tricks to play, but the speed is fine for the
test app of secchi depth.