src/testing: working on map polygons
The hope is that speed will be increased by grouping polygon plotting
into one call.  So far, the test code does not do any of the offpage
and other testing.  Also, it is not stitched in to the R code; to test
it, use lines in the comment block above the function.