tests/colormap.R gets very specific tests
For one test case, from issue 449, every single element in the colormap
is checked against a value that seemed OK to dk and cr.  Naturally,
this test will need to be changed if the desired outcome is altered,
but for now it is a protection against inadvertently changing the
colormap in future code alterations.