ctd: clean up some accessor names
potentialTemperature and absoluteSalinity removed because they served
no real function and were confusing.  Similarly, Theta was removed
because, although it makes sense for TEOS-10, it conflicts with the
function name swTheta() … besides, it’s a stupid abbreviation because
it requires very precise TeX typesetting (JPO even had to alter their
formatting rules to permit it).