fix to read.logger for when there are more names than data channels
Not exactly sure what is going on with this yet, but it seems like an
RSK file *can* be populated with more "names" than there are data
channels. This seems to occur because sometimes derived variables are
stored in the database, and if they aren't the names are anyway.

The problem was in the line:

names(data) <- names

which I changed to

names(data) <- names[1:dim(data)[2]]

so that it only assigns names to the channels that are actually in the database.