rgdal but docs broken; cannot plot ctd map
BUILD WARNINGS: The problem with the docs is that generics on [[ and
plot are said to be missing.

ERRORS (commented-out features): The plot(ctd, which="map") problem is
that within the ctd code, the call to plot the coastline seems to try
using generic plot() even though it is a coastline object.

SUMMARY: Something is messed up with the objects.  I do not know yet if
the problem is with including rgdal, but it's hard to extract that at
this stage.