only suggest rgdal, and now builds ok
Perhaps there was a problem with where in NAMESPACE I was mentioning
rgdal, or how it was mentioned. In any case, I prefer to just suggest
it, because many users won't want projections anyway, and the rgdal
startup message is really rather distracting to see every time oce is
loaded.