First commit adding Ruskin Rtext export to read.rsk
oceMagic update to recognize the Rtext files, but read.oce doens't
know about them yet.

To get read.rsk to work, the `type` arg has to be `txt`. The way the
file is read in is different from the code for the older format files
(e.g. it uses read.table), but should be clearer and is (I think)
robust.

Still need to add subsetting capabilities.