test: slot named flags
I coded this to see if it would cause large coding disruptions and
whether it would feel right. There were some initial coding disruptions
and I suspect others as well, but not too much to handle. As for the
“feeling”, well, I don’t like it. We have had 3 slots for a VERY long
time now, and have never discussed adding a fourth. That, alone,
suggests being cautious on making a new slot. But the other is that
this *really* seems to be metadata. Put another way, if flags are not
metadata, then what is? (I take “metadata” to mean “data about data”.)
Therefore, I plan to move flags from this new slot into metadata, and
am only checking in now to set the history.