mapPlot() longitude bug fixed (issue 766)
This was caused by the fact that a "longitudelim" in the 0 to 360
notation was passed by mapPlot() to mapGrid(), and the latter was trying
to draw lines in the -180 to 180 notation. I made a new function called
shiftLongitude() that maps back to the -180 to 180 notation. Since the
function might prove useful more generally, I exposed it to the user.
Later, it could get an argument telling it which system to prefer, but a
default of -180 to 180 is fine for now.