More robust fixed for duplicated channels in rsk files
i.e if there are three channels called temperature, the first will
remain as is, the second will be temperature2, and the third
temperature3. This should work for an unlimited number of duplicated
channels, even though in practice there will likely be two at the most
(except for some special loggers).