put T/C units in metadata$units, a list
This involved changing a lot of files, so we can expect new bugs, even
though I tried to make the accessor functions be robust across the
change. I rebuilt the ctd, rsk and section objects to be in this new
format, which should mean that objects created with the previous system
will still "work". However ...

EXPECT BUGS and please report them!