A codebase unification has changed *many* files.
1. All datasets have been updated to the new base `oce` class so that
all objects automatically have `metadata$units`. This is the sort of
thing that happens only when the base class changes, which is an
exceedingly rare thing. (I think this is the first change.)

2. I bit the bullet and renamed all `rval` objects as `res`. This will
eliminate the idiotic pattern I have of almost randomly selecting a
name.  I should have decided on day 1 how to name return values.

3. I think (although I've not checked) that *all* the objects now have
corresponding entries in the `create_data` directory. That does not mean
that the actual raw data are provided, because this directory is really
for the use of the developers, and they share some of the underlying
data privately.

4. All lists should contain 3 items, so this item will self-destruct in
15 seconds. The agency will disavow all knowledge of this item.