full-package code improvement
I finally bit the bullet and decided to address the many thousands of
warnings that resulted from the Rstudio ‘code diagnostics’. These were
mainly in two categories:

1. Use of global variables in functions. This was solved by making
those global variables into arguments to the functions — an easy fix in
the editor.

2. Unchecked missing values. This was a lot harder to solve, and I
worry a bit that some of the changes to the logic may come back to bite
us! There were two solutions. In some cases, I linked if() blocks and
removed early returns. This is simply a better coding style and things
should have been that way from the start. In other cases, I added
defaults to function  arguments. Usually the default was set to NULL,
so the missing(x) simply became is.null(x). In other cases it seemed to
be more sensible for the default to be NA.

None of the above changes should alter the response to user code … or,
put another way, any alterations will be considered bugs.

There is one thing that may affect user code. This is a pruning of one
of the choices for the ‘submethod’ argument of ctdTrim(). The new
situation is described in the documentation. It results because the old
code seems a bit useless to me and, frankly, I couldn’t really
understand either the old code or the old documentation. I noticed this
confusing code because I was fixing one of these code-diagnostic items,
for the submethod called bilinearA.

As of now, the whole codebase (a) checks without problem and passes the
code-diagnostic test without one single warning. Yes, reader, that’s
right: the diagnostic listing shrank for screenful after screenful (I
didn’t bother to estimate — hundreds? thousands?) to not. one. single.
warning. :-)