units but still build NOTE
I think I will need to 'protect' the units info, because the data() that
we supply has those units. That will make the output of
str(object@metadata$units) look really confusing. Certainly, I don't see
an advantage to just saving e.g. "degree C", which is (a) readible by
anyone and (b) not a problem for CRAN.

So, tomorrow on Jan 7 I will likely give up on this hours-old idea of
storing unicode characters for the degree symbol, for Greek mu, etc.

It's not at all clear to me why CRAN feels it must check that characters
in *datasets* must be ASCII. But ours is not to wonder why.