Finding different *.nod files
If there are different naming standards used at the moment, this might be a better solution, just looking for a .nod file in the folder (just in case new datasets are added and a different naming standard is used). As far as I can see there is always just 1 nod file in the data folder.

All the datasets I tested worked alright with this patch (tested on artic9, brador, h3o, HRglobal, ne_pac4, mwatl, sshelf).