broken state...
I am working near odf.R:283. I want to make this fcn return a list, just
like its cnv cousin. And I want units to be determined here, not in the
calling code; again, this mimics the cnv code. Then it will be easy to
mimic what the cnv code is doing with columns.