4th travis-ci attempt (speed)
Although attempt 3 worked, it was a 24 minute test. That's about a
factor of 4 or 5 slower than it was before the present scheme of
building a full rgdal system on travis-ci before testing oce. And what's
next? Today I fixed ncdf4 and rgdal, but tomorrow it could be some other
thing. I want to deepen my understanding of how to deal with the
shifting sand that is travis-ci. So, the present commit is an attempt to
install rgdal from ubuntu binary, which should be faster than building
it from source. I don't know if I have to also specify an rgdal library
... I guess we'll wait and see, fingers crossed!