new scheme to location adp ensemble pointers+
+ the timing for a Mb file is as follows. The first values are for the
old scheme, and the second for the new. Both times are very much smaller
than the time of decoding the information in the file, so I judge the
new code's speed to be entirely adequate, especially given that this new
scheme may provide the foundation for R code that can handle large files
(issue 1093).

   user  system elapsed
  0.013   0.001   0.014
   user  system elapsed
  0.052   0.017   0.069