improve setup of VMDAS storage+
+this came up in a test file I was looking at for the adp test suite. I
don't remember which file it was. The problem was that in the previous
scheme, storage got set up on the the test `i=1` but, for whatever
reason, that was not occuring. In any case, the code is clearer if we
simply set a flag, and then unset it when we create storage, and so that
is what this change does. Note that this is done also in the `dk`
branch, but I do not plan on merging `dk` into `develop` for a while, so
I did this directly. (And, in so doing, I found a second spot that
needed this.)