fix CTD fallback of subset,section-method
It looks to me as though this must not have actually worked before. The
code was assuming that the CTD data slot was a data.frame, not a list.
This tells me that it was never actually tested, because I am fairly
certain that the writing of subset,section-method was greatly predated
by the CTD transition from data.frame to list. Given this evidence, I
didn't bother opening an issue to explain this commit.