change 2 local time tests
1. test_local_adcp.R #ensembles is now different (correct, I think)

2. test_local_ctd.R give up on getting institution name from the silly
first line of headers.  (It turns out that the format of this line is
not fixed by WOCE convention, apart from the date. This means that there
is no way to know which letters are an institution abbreviation and
which are the initials in the analyst's name. Yes, you read that right.
Don't blame the messenger.)