address most warnings during devtools::test()
Some of these led to good changes in the code, but most were stating the
warnings we expect, from data that we know to be flawed (e.g. ODF data
with broken CRAT units). NOTE: it's not clear to me how different
languages will be handled, if the warning is translated but not the
test, but I think that's a matter for consideration by the authors of
the testhat package (and, given who they are, my guess is that they are
aware of this matter).