sfm_enu conversion more robust
There was a problem if the user had not supplied pitch, heading and roll
as vectors of length equal to the number of sampling times.  (In the
code I call the number of sampling times as "np", thinking "number of
profiles" as for 'adp' type, but I use the same variable name for 'adv'
type also.)

The problem is solved -- at the R level -- by lengthening heading, pitch
and roll as needed.

I also tweaked the C code in src/sfm_enu.c by making it expect all
vectors to be of the same length. Previously, the code was recycling
pitch, heading and roll if required, but this scheme was not being used
correctly in the R code (which proves it was a bad scheme) and it
entailed code duplication that is dangerous (e.g. a developer might
modify one block but forget to modify it's doppelganger).

I have built and tested oce but this will not really exercise this
change. The way to exercise the change is to use the R code from the Rmd
code that is mentioned in https://github.com/dankelley/oce/issues/1250
and to run it with valgrind, i.e.

    R -d valgrind --no-save < code.R

to see if allocation or access errors are reported.