read.echosounder() handles soundSpeed better
I noticed that the value being calculated from the S and T in the file
was not actually used. But the default that *was* used was a bit
problematic, because it was not specifying the eos, and not providing
lon and lat. I switched the default to use eos="unesco" ... and this is
not something to fret over, because we are just using a
picked-from-the-air value for S and T, anyhow.