mapPlot experiments
I was seeing a problem with parallels not closing, and that seemed to be because of some UHL-quenching code. Behaviour depends on par(mar), because this relates to marching around the perimeter of the plot box,  trying to invert the data to get ranges on lon and lat. That's fine if we are seeing just a portion of the earth, but it fails with views showing the earth and space around it. Maybe I should construct interior points, not just along the perimeter ... I'll play with this over the next few days in the 'dk' branch (which should not be used by anybody else).