update coastlineWorld base dataset
I don't know how much it differs, but I was checking that the link still
worked, and so I re-downloaded it.  It seems to plot OK, and the size is
similar (to a couple of percent) so I am confident that this is not
something wildly different from what it was before ... for example, we
still have country boundaries.