[[,tidem-method handles "frequency" etc
The accessor now handles "frequency", "amplitude", and "phase" directly.
This is also documented.  The main addition, for users, is that they
don't have to learn that frequency is called "freq" in these objects.
(Admittedly, 'freq' is a silly name, but it cannot be changed easily at
this time, because users may have written code that uses that name.)