make ?`initialize,ctd-method' work (PLAN...
... I may do this for all objects. I had been trying to do e.g.

.ctd <- setClass()

as Hadley Wickham suggests is better than `initialize,X-method`
(https://adv-r.hadley.nz/s4.html) but I just could not get that working,
for reasons that escape me.  Also, I could not see any way to get the
kind of documentation that would be easy for the user, so they could do

?.ctd

and see that the first arg is some hydrographic quantity, and not the
class name "ctd". After quite a bit of searching, in fact, I found
nothing really satisfactory for documenting S4 methods.  I think the
best plan might be to look at the bioconductor source, because
apparently that project is S4-based.