?'odf-class' improved+
+this is a test case for how I want to do accessor and slot docs. I
intend to try it out on ctd-class next. I don't have any particular rush
on this, and I'd prefer to do it a bit at a time, so I can continually
check that the wording seems OK. This is the beauty of templates.

Also, summary() now works even for two-row datasets that contain a time
row. (This was a problem in R/AllClass.R with forgetting to use a
drop=FALSE in a subset.)