plotProfile() handles units better (issue 1403)+
+ this should improve plot,ctd-method also, of course.  The problem came
up in fixing another issue in which axis labels for unitless things were
getting "[]" appended on the end, which looked like a weird cabalistic
box on the plot (as it does, in this commit message, with the font I'm
using in my terminal!).

I am reasonably assured that the fix is OK, from inspection of the "git
diff" output and from the 1403 test suite.