read.odf() handles multiple NULL_VALUE
And the test suite now has a test file. I may alter that, though,
because the file is quite large.  (Maybe I'll also shorten the file in
inst/extdata, also ... I don't want to bloat oce for this test, but I
made an actual file instead of setting up a temporary file, applying a
patch (how, on windows?), etc.