remove url for SBE processing manual
The link http://www.seabird.com/document/sbe-data-processing-manual has
been failing for several months now. I just spent some time looking
carefully on the SBE website to try to find this file, or something like
it, and have failed.  If I ever find the file, I'll reactivate the link,
but I am unwilling to continue to tolerate make-check errors on this.
The docs continue to display this URL, but they make note of the fact
that it's broken, and that several research groups are providing copies
of the documentation, so readers can actually find the material (perhaps
in abrogation of contract).