read.ad2cp() handles v,a,q better
If any of these is not present, the others can still be read. An
offset counter is used whilst traversing the data, and this (when
extended more) will permit reading the transformation matrix, which is
placed after optional elements.

Localized build-tests have also been added for v, a, and q.