ad2cp supports and tests interleavedBurst
Actually, I am basically only guessing on what is in this data-record
type, because the nortek docs are very hazy on this. (The string
"interleave" appears only once in the whole integrator-guide document!)

The test plot at the end of tests/testthat/test_ad2cp.R shows the
results for a private sample file. In that case, the interleaved burst
was for a single beam, perhaps a vertical one, but I really just do not
know, and my goal at this stage is merely to get code that can scan the
data, because that's the hard part ... we can figure out what the data
mean later, if a user can tell us something about the data.