ad2cp change echosounder code but...
... this is still pretty much a guess, based on nortek docs that are not
easy to understand. For example, do records of 'echosounder' type also
contain velocity, amplitude, and quality? And are those things just for
the central beam, or all of them?  I really think the answer is "no" --
that is, I doubt that echosounder records have anything but, well,
echsounder data. In any case, I think the code may work OK, because I
have tests that only construct (and later fill) the 'v', 'a' and 'q'
arrays if the record header indicates that such data are present.

I need data to test all of my assumptions, and of course to test the
code, which was created by a copy/paste/modify cycle that may well be
error-prone.