read.adp.ad2cp() reads mag declination
I am inferring it from DECL in the header, since the letters 'decl' do
not come up in the docs anywhere except in that definition. But I am
wondering a bit about this, because the three test files I have all have
DECL=0.0, which seems an odd coincidence since they are from different
research teams and institutions.  Maybe users are doing what I'd do,
though -- setting it to 0.0 at the setup stage, planning to use it later
on, during data processing.