plot(as.ctd(x)) handles multistation data (1484)
This came up for glider work, but it could be good for sections as well,
e.g.

plot(as.ctd(read.section("SECFILE")), type="p")

would make a nice summary plot. We need the type="p" to avoid lines
connecting the tops and bottoms of station sequence in the profile plots
and I guess the TS also.