read.rsk() handles patm properly
I think this works now, both on my actual file and on the newly-added
tests in the test suit. (That test suite requires a rsk file that is not
built into oce, so it won't get run by external checkers that work with
the package, but it should work within the git repository, because the
repository holds this extra .rsk file.)

Fixes issue #1491.

Richard III, Act 1 Scene 1

> Now is the winter of our discontent
> Made glorious summer by this sun of York;
> And all the clouds that lour’d upon our house
> In the deep bosom of the ocean buried.