binApply2D tests
I want to check more carefully, though, and may very well just use a
double loop, which will be clear to any reader. Also, there are lots of
minor changes here:

1. Many files have dontrun switched to dontest.

2. I reformatted both dontrun and donttest to remove leading blanks.
This is to make it easier to find them with grep.  While I was in these
R files, I added some blank roxygen2 lines, to make the source easier to
read.

3. I altered the tide-related family (since e.g. webtide is not related
to tidem, yet it is tide-related).