binApply2D doc example fixed (issue 1493)
The code was somewhat opaque, involving split and cut, and so I have
replaced it with simple loops over x and y. Actually, the old method --
now devoid of this bug -- is still available in the code, as method=1,
starting at R/misc.R line 890. I have checked that method=1 and method=2
give the same results, using the test in tests/testthat/test_misc.R, and
I suppose I can switch to method=1 if the present scheme proves to be
slow on large matrices. However, indexed loops are now fast in R, and
opaque code is dangerous (as this issue proves!), so I am inclined to
leave this simpler method=2 code as the default.