clean some build-test errors in docs
These were previously hidden in dontrun{} blocks, but now they are in
donttest{} blocks, so we can test them. Soon, if rgdal gains
functionality on i396/windows, we can even remove the donttest blocks,
except for any that burn up so much time that they run into a problem
with CRAN policy.