rgdal kludge re version number
rgdal::project() now has an arg to handle NA on i386/windows, but the
older versions of rgdal lack that, so we must check the version number
and have two ways of working (one of which will fail on i386/windows,
unless the user updates rgdal).

A way to avoid forcing users to upgrade rgdal would be to temporarily
replace these NA values with something that is okay, e.g. perhaps (0,0)
on the forward projection, and then replace with NA after the
projection. This will ought to yield the same behaviour on a range of
machines and versions.