read.argo attempts to read all 'extra' info into metadata
This change is in response to

https://github.com/dankelley/oce/issues/1520

I do not have the test file that inspired this issue report, so instead
I tested it with

a <- read.argo("create_data/argo/6900388_prof.nc")

Note that this produces 12 warnings, because the ncdf4::ncvar_get()
function has problems with dimensionality.

I think we need to get a test with the data that @richardsc is actually
trying to read.

Note that I do not rename these things -- they are stored with names
LIKE_THIS, without converting to likeThis.

PS. I think the docs for read.argo() probably will need an edit on this
matter, since it is new behaviour. But I will wait until I get some
confirmation from @richardsc on whether what I am doing is useful.