1. I have updated the "develop" branch to handle this. The change caused some ripple effects, and below I am putting some detailed notes, which I hope other developers will consider reading, as a check on whether I've gone off the rail. I'm numbering the paragraphs for reference, if further discussion is needed. Since paragraph 4 relates to a change from old behaviour, I ask that co-developers note that one, in particular ... I can change back to the old behaviour, but I think we ought instead to do as I have it now. (Basically, I always would prefer to compute density from the others, to avoid problems where maybe someone stored the file with density digits that don't make sense, given the S, T and p digits. This is a 10+ year old decision, but for some reason I didn't follow it in the ODF case.)
2. The changes involve docs, code, and tests. The docs now have a 5-item list
of how things are looked up (with this new [[item,"data"]] and
[[item,"metadata"]] aspect being number 5 in that list.)  I redid the logic of
how it looks for things, and this is now explained in a lot of detail in the
output from ?"[[,oce-method". I also had to remove the specialized operation of
"[[,ctd-method" with regard to "sigmaTheta", "sigma0" and "spice", because
these things are now handled at the deeper "[[,oce-method" level. My initial
coding caused problems in the test suite, where we accessg these calculated
things for a newly-constructed oce object...

3. So, it proved necessary to go into the R/sw.R code and modify the handling
of args.  There are lots of special cases that need to be checked (including
new checks I added throughout this work). In particular -- and @richardsc will
get a kick out of this -- I removed the "lookWithin" madness within swSigm0(),
swSigmaTheta() and swTheta() ... and possibly other functions. I was not too
worried about changing this code because of all the tests we have in the
tests/testthat/*.R files. Frankly, I am not really seeing the purpose in
retaining the lookWithin methodology at all, and so I may remove it over the
weekend ... of course, ensuring that old tests still work and adding new tests.

4. Why do I think it's smart to remove lookWithin? Well, I only really
understood what the heck I was doing way back when I first started that (for
gsw, I think), and now when I look at how it's used in oce, I am a bit
mystified. This not a feature of code that can be maintained!

5. I am quite assured that the changes all work OK, because the test suite
touches on them. I added a few new tests to the suite, for this new
[[item,"data"]] and [[item,"metadata"]] scheme, but also for other
accessor-related functionality. A lot of this came from the fact that we can do
e.g. x[["sigmaTheta"]] and swSigmaTheta(x), and my plan for this afternoon is
to redo all the tests on my home machine, with global "oceEOS" set to "unesco",
as opposed to the "gsw" tests that I used this morning at work.

5. I realized that the handling of odf data was odd, with regard to sigmaTheta.
Below is some new content in tests/testthat/test_odf.R. The change is in the
d[["sigmaTheta"]] part.

   ## Test sigmaTheta in two ways. Using [["sigmaTheta"]] will *compute*
   ## the value (so it will be NA when S, T or p is NA), but
   ## [["SIGP_01"]] access the *supplied* value, so it is NA at line 18.
   ## NOTE: the behaviour for the [["sigmaTheta"]] case was to return the
   ## same as the [["SIGP_01"]] case until 2019 May 17, when some changes
   ## got made for issue 1554 (see
   ## https://github.com/dankelley/oce/issues/1554), whereupon I realized
   ## that the old behaviour made ODF different from other oce data
   ## types, which is not sensible.
   badSTp <- which(is.na(d@data$salinity)|is.na(d@data$temperature)|is.na(d@data$pressure))
   expect_equal(badSTP, which(is.na(d[["sigmaTheta"]]))) # computed
   expect_equal(18, which(is.na(d[["SIGP_01"]]))) # original data
   expect_equal(19, which(is.na(d[["sigmaThetaFlag"]])))