fix magdef.f wrt a fortran warning
The problem is in doing e.g.
```
       11 = 0.02(date - 1995.0)
```

which assigns to (an implicitly typed) INT4, from a REAL8.  Wrapping the
RHS in int() makes the warning go away. But if we actually *need* REAL8
in the RHS, the fix might be more sophisticated. However, all we are
doing is taking the integer part of 0.2 times the number of years since
1900, and surely we don't need REAL8 for that!

We have several tests on magnetic fields in `tests/testthat/test_misc.R`
and so I see no reason to leave issue
https://github.com/dankelley/oce/issues/1560 open.

https://github.com/dankelley/oce/issues/1560