sontek website changed from a proper URL to a phrase
This is because sontek now has a popup that demands the reader agrees to
privacy agreement with respect to somethig caled Xylem. Obviously, the
automated CRAN build/check engines are not going to fake a user click
for that.

Perhaps this means that sontek lose the chance for some business, since
oce will no longer provide links to their site. Tough.