?read.odf no longer states when URL last checked
The new policy is to use rhub and other tests to find broken links, and
to change the markup from \url{} to \code{} if a failure is persistent.
Otherwise, for consistency, we'd basically have to update
last-good-check dates all through the docs. I don't see the change to
\code{} as being a big problem for users, because they can easily copy
the text and paste it into a web browser.