read.adp.rdi() trims time (if needed)
This addresses issue https://github.com/dankelley/oce/issues/1579 by
trimming the 'time' vector if its length exceeds the length of
`profileStart'. (The latter is used in constructing the other vector and
array data.)

A real-time message is printed when this is done, and a warning is
issued as well.

I do not know whether this problem is particular to this file, but the
fix is very simple and I am not worried about ripple effects for other
files.