trying to fix issue 1583
Actually, the issue as stated there is ok now (I think) but plotting is
not working. Is Barnes filling in the matrix?  (I have new debug input
on that.)  But I don't even see in the plot,section-method code where I
am assigning the zz matrix ... I plan to look into this tomorrow
morning.