tide functions now skip unknown constituents
This stemmed from
https://github.com/dankelley/oce/issues/1592
where a constituent named 'M1' is not handled. I've learnend that it is
a real constituent, but it is not in the Foreman (1987) list, and it's
not handled by T_TIDE, so it makes sense for oce to not handle it
either.

In writing code to detect unknown constituents, I also cleaned up
as.tidem(), since that was the function that made me twig to 'M2'.

I also did some housekeeping, changing all the docs in R/tides.R to
markdown format, from the older Rd format.  (Since markdown is easier to
maintain than Rd, my intention is to convert to markdown whenever I'm
altering a function, and need to alter the docs, or read the docs to
understand what the function is meant to do.)